+++ sdk/resourcemanager/resourcegraph/armresourcegraph/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 0.6.0 (2022-05-18)
+### Breaking Changes
+
+- Function `ErrorDetails.MarshalJSON` has been removed
+- Function `Table.MarshalJSON` has been removed
+- Function `Error.MarshalJSON` has been removed
+- Function `FacetError.MarshalJSON` has been removed
+- Function `FacetResult.MarshalJSON` has been removed
+- Function `QueryResponse.MarshalJSON` has been removed
+- Function `OperationListResult.MarshalJSON` has been removed
+
+
 ## 0.5.0 (2022-04-18)
 ### Breaking Changes